+++ changelog.md
@@ -462,3 +462,9 @@
     // ver 0.10.28 @ 2025-09-06
     //    kino: fix infinite running (bug of v0.10.26)
     //    util: fix log causing error
+    // ver 0.10.30 @ 2025-11-14
+    //    jw: fix minor scraping logic
+    //    watcha: fix minor logic in edit() on large div
+    // ver 0.11.0 @ 2025-11-14
+    //    internal: minor wrap-up
+    //    get rating (which is not on jw) directly from imdb via opening tab (rare case)